requirements.txt
acdh-django-sparql==0.0.1
bleach==3.0.2
certifi==2018.8.24
chardet==3.0.4
Django==2.1.2
django-autocomplete-light==3.3.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-leaflet==0.24.0
django-reversion==3.0.0
django-tables2==2.0.1
djangorestframework==3.8.2
docutils==0.14
future==0.16.0
idna==2.7
isodate==0.6.0
lxml==4.2.5
numpy==1.15.2
pandas==0.23.4
pkginfo==1.4.2
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
rdflib==4.2.2
readme-renderer==23.0
requests==2.19.1
requests-toolbelt==0.8.0
six==1.11.0
SPARQLWrapper==1.8.2
tqdm==4.28.1
twine==1.12.1
urllib3==1.23
webencodings==0.5.1
wincertstore==0.2
